% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_interval.R
\name{build_interval}
\alias{build_interval}
\title{Construct Interval Indices}
\usage{
build_interval(n)
}
\arguments{
\item{n}{outer boundary of interval}
}
\description{
Functions to create interval indices
}
