% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIndexPairs.R
\name{GetIndexPairs}
\alias{GetIndexPairs}
\title{Pair an Index for Performance Comparisons}
\usage{
GetIndexPairs(n, asMatrix = FALSE)
}
\arguments{
\item{n}{outer boundary of interval}

\item{asMatrix}{boolean flag where TRUE returns Matrix, FALSE for list}
}
\description{
Pair an Index for Performance Comparisons
}
