% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIndexPairs.R
\name{index_pairs}
\alias{index_pairs}
\alias{GetIndexPairs}
\alias{converge_to_mid}
\title{Pair an Index for Performance Comparisons}
\usage{
GetIndexPairs(n, asMatrix = FALSE)

converge_to_mid(n)
}
\arguments{
\item{n}{outer boundary of interval}

\item{asMatrix}{boolean flag where TRUE returns Matrix, FALSE for list}
}
\description{
Pair an Index for Performance Comparisons
}
\section{Functions}{
\itemize{
\item \code{GetIndexPairs}: TBD

\item \code{converge_to_mid}: TBD
}}

