% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigeon_maker.R
\name{index_pigeon}
\alias{index_pigeon}
\title{Pigeonhole Maker}
\usage{
index_pigeon(total_count, chunk_size, f = I, ...)
}
\arguments{
\item{total_count}{TBD}

\item{chunk_size}{TBD}

\item{f}{TBD}

\item{...}{TBD}
}
\description{
See examples
}
\examples{
VEC <- paste0("x", 1:1000)
f <- function(ind) VEC[ind]
index_pigeon(1000, 100, f, VEC)
index_pigeon(1000, 100, f, VEC, simplify = FALSE)

}
